# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BootstrapInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
}

export interface SessionInput {
  /** @minLength 8 */
  sessionId: string;
}

export type RuntimeStatusState = typeof RuntimeStatusState[keyof typeof RuntimeStatusState];


export const RuntimeStatusState = {
  offline: 'offline',
  waiting: 'waiting',
  connected: 'connected',
  busy: 'busy',
  error: 'error',
} as const;

export interface RuntimeStatus {
  state: RuntimeStatusState;
  /** @nullable */
  sessionId: string | null;
  /** @nullable */
  label: string | null;
  /** @nullable */
  connectedAt: string | null;
  /** @nullable */
  lastSeenAt: string | null;
  queuedCommands: number;
  /** @nullable */
  pythonVersion: string | null;
}

export interface BootstrapResponse {
  sessionId: string;
  token: string;
  connectorCode: string;
  expiresAt: string;
}

export interface ConnectInput {
  sessionId: string;
  token: string;
  /** @maxLength 120 */
  runtimeName?: string;
  /** @maxLength 80 */
  pythonVersion?: string;
}

export interface ExecuteInput {
  sessionId: string;
  /**
     * @minLength 1
     * @maxLength 30000
     */
  code: string;
  /** @nullable */
  description?: string | null;
}

export interface ExecutionResponse {
  accepted: boolean;
  commandId: string;
  message: string;
}

export type RuntimeEventType = typeof RuntimeEventType[keyof typeof RuntimeEventType];


export const RuntimeEventType = {
  status: 'status',
  stdout: 'stdout',
  stderr: 'stderr',
  result: 'result',
  error: 'error',
  command: 'command',
  system: 'system',
} as const;

export interface RuntimeEvent {
  id: string;
  type: RuntimeEventType;
  message: string;
  /** @nullable */
  payload: string | null;
  createdAt: string;
}

export interface RuntimeEventsResponse {
  events: RuntimeEvent[];
  cursor: string;
}

export type ColabEventInputType = typeof ColabEventInputType[keyof typeof ColabEventInputType];


export const ColabEventInputType = {
  status: 'status',
  stdout: 'stdout',
  stderr: 'stderr',
  result: 'result',
  error: 'error',
  system: 'system',
} as const;

export interface ColabEventInput {
  sessionId: string;
  token: string;
  type: ColabEventInputType;
  /** @maxLength 30000 */
  message: string;
  /** @nullable */
  payload?: string | null;
  /** @nullable */
  commandId?: string | null;
}

export interface ColabCommand {
  id: string;
  code: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
}

export interface ColabCommandsResponse {
  commands: ColabCommand[];
}

export interface AcceptedResponse {
  accepted: boolean;
  message: string;
}

export type AssistantInputProvider = typeof AssistantInputProvider[keyof typeof AssistantInputProvider];


export const AssistantInputProvider = {
  openai: 'openai',
  gemini: 'gemini',
  anthropic: 'anthropic',
  openrouter: 'openrouter',
  custom: 'custom',
} as const;

export interface AssistantInput {
  /** @nullable */
  sessionId?: string | null;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  message: string;
  provider: AssistantInputProvider;
  /**
     * @minLength 1
     * @maxLength 500
     */
  apiKey: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  model: string;
  execute: boolean;
}

export interface AssistantResponse {
  reply: string;
  /** @nullable */
  code: string | null;
  /** @nullable */
  commandId: string | null;
  provider: string;
  model: string;
}

export type GetRuntimeEventsParams = {
cursor?: string;
sessionId: string;
};

export type GetColabCommandsParams = {
sessionId: string;
token: string;
};