# lib/api-zod/src/generated/types/assistantInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantInputProvider } from './assistantInputProvider';

export interface AssistantInput {
  /** @nullable */
  sessionId?: string | null;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  message: string;
  provider: AssistantInputProvider;
  /**
     * @minLength 1
     * @maxLength 500
     */
  apiKey: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  model: string;
  execute: boolean;
}