# lib/api-zod/src/generated/types/assistantResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantResponse {
  reply: string;
  /** @nullable */
  code: string | null;
  /** @nullable */
  commandId: string | null;
  provider: string;
  model: string;
}