# lib/api-zod/src/generated/types/colabEventInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ColabEventInputType } from './colabEventInputType';

export interface ColabEventInput {
  sessionId: string;
  token: string;
  type: ColabEventInputType;
  /** @maxLength 30000 */
  message: string;
  /** @nullable */
  payload?: string | null;
  /** @nullable */
  commandId?: string | null;
}