# lib/api-zod/src/generated/types/connectInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectInput {
  sessionId: string;
  token: string;
  /** @maxLength 120 */
  runtimeName?: string;
  /** @maxLength 80 */
  pythonVersion?: string;
}