# lib/api-zod/src/generated/types/runtimeEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RuntimeEventType } from './runtimeEventType';

export interface RuntimeEvent {
  id: string;
  type: RuntimeEventType;
  message: string;
  /** @nullable */
  payload: string | null;
  createdAt: string;
}