# lib/api-zod/src/generated/types/runtimeEventsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RuntimeEvent } from './runtimeEvent';

export interface RuntimeEventsResponse {
  events: RuntimeEvent[];
  cursor: string;
}