# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  /** @minLength 8 */
  sessionId: string;
}